[CI] Use uncrustify to check compliance with code formatting rules
This patch introduces a new CI pipeline for checking whether new code
complies with the Dune formatting rules.  For this, it runs the
uncrustify tool over all source files of the module and checks whether
that leads to any changes.

The configuration file for uncrustify is the one advertised on the
webpage that discusses the coding style for Dune

  https://dune-project.org/dev/codingstyle/