Add option documentation to the uncrustify config file
The uncrustify tool is able to write documentation for each of the
many options it supports.  This commit imports this documentation
for all options that appear in the Dune uncrustify configuration file.

It also reorders the options to the same ordering used by uncrustify
when writing a default option file itself.  That way, it should be
easier to compare the Dune configuration with the default configuration.