Make FieldTraits less greedy
The idea of the `FieldTraits` class is to provide
the `field_type` and `real_type` of various types
representing numbers, vectors, ... .
So far there was a default implementation
```cpp
template<class T>
struct FieldTraits
{
  typedef T field_type;
  typedef T real_type;
};
```
This does not make much sense, for types that do not
represent numbers. The patch ensures that the default
implementation does only provide the typedefs if
`Dune::IsNumber<T>` is true and that it is empty
otherwise.

This also fixes a seemingly unrelated issue: If you
defined `operator*(T,T)` and `dot(T,T)` for your own
type (or concept) `T`, then `dot` is ambiguous,
because there's a very greedy implementation in
`dotproduct.hh`.