Extend transpose(...) utility function to store passed r-value references
Internally this patch changes the storage and template parameter of
`Impl::TransposedMatrixWrapper` such that the passed matrix is
always stored by value. By also adding support to pass and store
a `std::reference_wrapper` this allows to store references and values.

The interface function `transpose(matrix)` now either stores
`std::cref(matrix)` or `matrix` in the wrapper depending on whether
`matrix` is an l-value or r-value reference.