Remove cpp definitions resolved in downstream
The motivation of this change is to remove the dependency of config file between modules. Before this change, these cpp definitions where resolved in the downstream module consuming dune-common. With this change, the config file is fully resolvable locally and we can transition to a non dune specific build system.