Allow equality comparison of TransformedRangeIterator classes with same base iterator
So far comparision was only provided for identical `TransformedRangeIterator`
type. This is generalized to allow equality/inequlity comparison of
different types as long as the base iterator is the same. This is reasonable,
since comparisons anyway rely on the base type only. The patch is needed
if we want to allow to generate `TransformedRangeIterator`s only with
lambdas, since each lambda has a different type.

The latter functionality is tested and demonstrated in the respective test.
We may later want to provide a public factory method for creating such
iterators (which should probably stay in `Dune::Impl::` themselves).