Add BUILD_INTERFACE to paths registered on the package registry
Testing on the dumux pipeline showed that the finding the dune package was adding the include directories as raw strings in the package registry. This caused installed targets linked against all dune packages to contain build directories even after being installed. That's obviously wrong as build directories may not exist when the packages are installed. The offending line is already removed in master, but not for release 2.10 (see https://gitlab.dune-project.org/core/dune-common/-/merge_requests/1457). So this changes adds the `BUILD_INTERFACE` to fix the problem, and an opt-in in case someone was relying on the previous behavior.

Note that raw string in the registered include directories are technically a ~bug. However, in favor of stability wrt 2.10 release we keep the current behavior and add a flag (`DUNE_REGISTER_DUNE_DEPENDENCIES_WITH_BUILD_INTERFACE`) to enable the `BUILD_INTERFACE` generator expression when linking targets that will be exported.