[cleanup] Extract transform method from TransformationRangeIteratorTraits
This method is used to implement the different invocation modes.
It originally ended up in the traits class, because it is needed
here to deduce the reference type. However, an actual implementation
should not be part of the traits class.

The current patch fixes this flaw by extracting the method as
templated global function `invokeWithIterator`. Since the old
and new code live in `Dune::Impl::`, this change is not problematic.