Add filteredRangeView() utility
This provides a range view that filters another range using
a predicate. In contrast to `std::ranges::filter_view` it
does not cache the begin iterator and thus does not satisfy
the 'amortized O(1)' requirement of the range concept.
This decision is deliberate because it avoids many potential
pitfalls (e.g. that `begin()` is not `const` and not thread-safe)